Test/requirements.txt
fastapi
uvicorn
python-dotenv
speechrecognition
pyttsx3
jinja2
